requirements.txt
streamlit
pandas
matplotlib.pyplot
re
os
seaborn
warnings
plotly.express
plotly.graph_objects
plotly.io
numpy
spacy
en_core_web_sm